fastapi/pyproject.toml
[tool.poetry]
name = "fastapi--app"
version = "0.1.0"
description = "fastapi project for assignment 2"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
streamlit-option-menu = "^0.3.13"
openai = "^1.47.0"
pymssql = "^2.3.1"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
python-docx = "^1.1.2"
python-pptx = "^1.0.2"
datasets = "^3.0.0"
plotly = "^5.24.1"
streamlit-plotly-events = "^0.0.6"
uvicorn = "^0.31.0"
pydantic = "^2.9.2"
fastapi = "^0.115.0"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
